# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Banca Empresas API
  version: 0.2.0
  description: |
    API del pipeline de venta de Tarjeta de Crédito **LATAM Business** para Banca Empresas
    (Banco de Bogotá).

    ## Endpoints expuestos actualmente

    | Método | Ruta | Descripción |
    |--------|------|-------------|
    | `GET` | `/health` | Health check del servicio |
    | `POST` | `/api/power-apps/submit` | Simulador de Power App con comprobación de campos |
    | `GET` | `/api/sales-calls/health` | Health check de la feature de llamadas |
    | `POST` | `/api/sales-calls/calls` | Inicia una llamada de venta vía Fonema.ia |
    | `GET` | `/api/sales-calls/calls` | Lista las llamadas registradas |
    | `GET` | `/api/sales-calls/calls/{id}` | Detalle de una llamada (transcript, análisis) |
    | `GET` | `/api/sales-calls/calls/{id}/recording` | Proxy de la grabación de audio (con Range) |
    | `GET` | `/api/sales-calls/calls/{id}/handoff` | Prefill para la Power App de una llamada calificada |
    | `POST` | `/api/sales-calls/batches` | Crear campaña (batch calling) con política de pacing |
    | `GET` | `/api/sales-calls/batches` | Listar campañas |
    | `GET` | `/api/sales-calls/batches/{id}` | Progreso agregado de una campaña (polling) |
    | `GET` | `/api/sales-calls/batches/{id}/items` | Items (llamadas) de una campaña |
    | `POST` | `/api/sales-calls/batches/{id}/{action}` | Pausar / reanudar / cancelar una campaña |
...

...
    se documentarán aquí conforme se expongan.

...

...

...

...

...

...
servers:
...
paths:
...
  /api/activation-follow-up/cases:
    get:
      tags: [Activation Follow-up]
      summary: Listar casos de seguimiento de uso
      description: |
        Casos de tarjetas entregadas con campos calculados: días sin uso, fase
        (`al_dia` <30 / `mes_1` 30-59 / `mes_2` 60-89 / `mes_3` ≥90), riesgo de
        inactivación (≥75 días) y próxima llamada estimada.
      operationId: listFollowUpCases
      responses:
        '200':
          description: Lista de casos de seguimiento
          content:
            application/json:
              schema:
                type: object
                required: [total, casos]
                properties:
                  total:
                    type: integer
                  casos:
                    type: array
                    items:
                      $ref: '#/components/schemas/FollowUpCase'
    post:
      tags: [Activation Follow-up]
      summary: Marcar entrega de la TC finalizada (check del punto 5 del portafolio)
      description: |
        Idempotente por cliente. La PRIMERA vez crea el caso de seguimiento,
        dispara la llamada de felicitación (agente Fonema de seguimiento,
        `tipo_llamada=felicitacion`) y avanza el pipeline a `activation_follow_up`.
        Las siguientes veces devuelve el caso existente sin re-llamar.
      operationId: finalizeDelivery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FinalizeDeliveryRequest'
            example:
              clienteId: '9001234567'
              nombre: 'EMPRESA DEMO S.A.S'
              telefono: '3005551234'
              correo: 'contacto@empresademo.example'
      responses:
        '201':
          description: Caso creado (primera vez); indica si la felicitación se inició
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinalizeDeliveryResponse'
        '200':
          description: El caso ya existía (re-check); no se vuelve a llamar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinalizeDeliveryResponse'
        '400':
          description: Payload inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: INTERNAL_ERROR
            message:
              type: string
              example: An unexpected error occurred
...
    FinalizeDeliveryRequest:
      type: object
      required: [clienteId]
      properties:
        clienteId:
          type: string
          pattern: '^\d{4,15}$'
          description: NIT de la empresa (lead del pipeline)
          example: '9001234567'
        nombre:
          type: string
          maxLength: 160
          example: 'EMPRESA DEMO S.A.S'
        telefono:
          type: string
          description: Teléfono de contacto (se normaliza a E.164 +57...)
          example: '3005551234'
        correo:
          type: string
          format: email
          example: 'contacto@empresademo.example'
    FinalizeDeliveryResponse:
      type: object
      required: [caso, llamadaFelicitacionIniciada, yaExistia]
      properties:
        caso:
          $ref: '#/components/schemas/FollowUpCase'
        llamadaFelicitacionIniciada:
          type: boolean
          description: true solo la primera vez, si la llamada Fonema se pudo iniciar
        yaExistia:
          type: boolean
    FollowUpCase:
      type: object
      description: Caso de seguimiento de uso de una TC entregada, con campos calculados
      required: [id, clienteId, deliveredAt, lastUsedAt, reminderCount, diasSinUso, fase]
      properties:
        id: { type: string, format: uuid }
        clienteId: { type: string, example: '9001234567' }
        caseId: { type: string, format: uuid, nullable: true }
        clienteNombre: { type: string, nullable: true, example: 'EMPRESA DEMO S.A.S' }
        telefono: { type: string, nullable: true, example: '3005551234' }
        correo: { type: string, nullable: true, example: 'contacto@empresademo.example' }
        deliveredAt: { type: string, format: date-time }
        congratulatedAt: { type: string, format: date-time, nullable: true }
        congratulationCallId: { type: string, nullable: true }
        lastUsedAt: { type: string, format: date-time }
        lastReminderAt: { type: string, format: date-time, nullable: true }
        reminderCount: { type: integer, example: 1 }
        diasSinUso: { type: integer, example: 34 }
        fase:
          type: string
          enum: [al_dia, mes_1, mes_2, mes_3]
          description: al_dia <30 días / mes_1 30-59 / mes_2 60-89 / mes_3 ≥90 (inactivación)
        riesgoInactivacion: { type: boolean, description: 'true desde el día 75' }
        diasParaInactivacion: { type: integer, example: 56 }
        proximaLlamadaEstimada: { type: string, format: date-time, nullable: true }
...
  securitySchemes:
    cronSecret:
      type: http
      scheme: bearer
      description: Bearer $CRON_SECRET inyectado por Vercel Cron.